# app/src/main/resources/openapi.yaml
openapi: 3.0.2
info:
  contact:
    email: openapi@inss.ch
  description: "Forked and modified by International Network Support & Service - Glas.\n\
    Modification are -> \n - removed xml and x-www-form-urlencoded (leaving only application/json).\
    \ \n - Adding examples to paramters. \n - Validated with openapi-style-validator-cli.\
    \ \nThis is a sample Pet Store Server based on the OpenAPI 3.0 specification.\
    \  You can find out more about\nSwagger at [http://swagger.io](http://swagger.io).\
    \ In the third iteration of the pet store, we've switched to the design first\
    \ approach!\nYou can now help us improve the API whether it's by making changes\
    \ to the definition itself or to the code.\nThat way, with time, we can improve\
    \ the API in general, and expose some of the new features in OAS3.\n\nSome useful\
    \ links:\n- [The Pet Store repository](https://github.com/swagger-api/swagger-petstore)\n\
    - [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)\n\
    - [OpenAPI samples from inss.ch](https://gitlab.com/networkinss/sampleopenapicollection)"
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://inss.ch
  title: Swagger Petstore - OpenAPI 3.0
  version: 1.1.0
...
servers:
- description: Localhost with port 8080.
  url: http://localhost:8080/api/v3
...
paths:
  /pet:
    post:
      description: Add a new pet to the store
      operationId: addPet
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pet'
        description: Create a new pet in the store
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
          description: Successful operation
        "405":
          description: Invalid input
      security:
      - petstore_auth:
        - write:pets
        - read:pets
      summary: Add a new pet to the store
      tags:
      - pet
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: pet
    put:
      description: Update an existing pet by Id
      operationId: updatePet
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pet'
        description: Update an existent pet in the store
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
          description: Successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Pet not found
        "405":
          description: Validation exception
      security:
      - petstore_auth:
        - write:pets
        - read:pets
      summary: Update an existing pet
      tags:
      - pet
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: pet
  /pet/findbystatus:
    get:
      description: Multiple status values can be provided with comma separated strings
      operationId: findPetsByStatus
      parameters:
      - description: Status values that need to be considered for filter
        explode: true
        in: query
        name: status
        required: false
        schema:
          default: available
          enum:
          - available
          - pending
          - sold
          example: available
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Pet'
                type: array
          description: successful operation
        "400":
          description: Invalid status value
      security:
      - petstore_auth:
        - write:pets
        - read:pets
      summary: Finds Pets by status
      tags:
      - pet
      x-accepts: application/json
      x-tags:
      - tag: pet
  /pet/findbytags:
    get:
      description: "Multiple tags can be provided with comma separated strings. Use\
        \ tag1, tag2, tag3 for testing."
      operationId: findPetsByTags
      parameters:
      - description: Tags to filter by
        explode: true
        in: query
        name: tags
        required: false
        schema:
          example:
          - tags1
          - tags2
          items:
            type: string
          type: array
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Pet'
                type: array
          description: successful operation
        "400":
          description: Invalid tag value
      security:
      - petstore_auth:
        - write:pets
        - read:pets
      summary: Finds Pets by tags
      tags:
      - pet
      x-accepts: application/json
      x-tags:
      - tag: pet
  /pet/{petId}:
    delete:
      description: Deletes a pet.
      operationId: deletePet
      parameters:
      - description: Api key.
        explode: false
        in: header
        name: API_KEY
        required: false
        schema:
          example: apikeyvalue1
          type: string
        style: simple
      - description: Pet id to delete
        explode: false
        in: path
        name: petId
        required: true
        schema:
          example: 99
          format: int64
          type: integer
        style: simple
      responses:
        "400":
          description: Invalid pet value
      security:
      - petstore_auth:
        - write:pets
        - read:pets
      summary: Deletes a pet
      tags:
      - pet
      x-accepts: application/json
      x-tags:
      - tag: pet
    get:
      description: Returns a single pet
      operationId: getPetById
      parameters:
      - description: ID of pet to return
        explode: false
        in: path
        name: petId
        required: true
        schema:
          example: 99
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
          description: successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Pet not found
      security:
      - api_key: []
      - petstore_auth:
        - write:pets
        - read:pets
      summary: Find pet by ID
      tags:
      - pet
      x-accepts: application/json
      x-tags:
      - tag: pet
    post:
      description: Updates a pet in the store with form data.
      operationId: updatePetWithForm
      parameters:
      - description: ID of pet that needs to be updated
        explode: false
        in: path
        name: petId
        required: true
        schema:
          example: 99
          format: int64
          type: integer
        style: simple
      - description: Name of pet that needs to be updated
        explode: true
        in: query
        name: name
        required: false
        schema:
          example: dog
          type: string
        style: form
      - description: Status of pet that needs to be updated
        explode: true
        in: query
        name: status
        required: false
        schema:
          example: in progress.
          type: string
        style: form
      responses:
        "405":
          description: Invalid input
      security:
      - petstore_auth:
        - write:pets
        - read:pets
      summary: Updates a pet in the store with form data
      tags:
      - pet
      x-accepts: application/json
      x-tags:
      - tag: pet
  /pet/{petId}/uploadimage:
    post:
      description: Uploads an image.
      operationId: uploadFile
      parameters:
      - description: ID of pet to update
        explode: false
        in: path
        name: petId
        required: true
        schema:
          example: 99
          format: int64
          type: integer
        style: simple
      - description: Additional Metadata
        explode: true
        in: query
        name: additionalMetadata
        required: false
        schema:
          example: metadata
          type: string
        style: form
      requestBody:
        content:
          application/octet-stream:
            schema:
              format: binary
              type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
          description: successful operation
      security:
      - petstore_auth:
        - write:pets
        - read:pets
      summary: uploads an image
      tags:
      - pet
      x-content-type: application/octet-stream
      x-accepts: application/json
      x-tags:
      - tag: pet
...
components:
...
  schemas:
...
    Category:
      example:
        name: Dogs
        id: 1
      properties:
        id:
          description: Id
          example: 1
          format: int64
          type: integer
        name:
          description: Name of category.
          example: Dogs
          type: string
      type: object
...
    Tag:
      example:
        name: tag1
        id: 1
      properties:
        id:
          description: Id
          example: 1
          format: int64
          type: integer
        name:
          description: Name of tag.
          example: tag1
          type: string
      type: object
    Pet:
      example:
        photoUrls:
        - http://example.com/photo.png
        - http://example.com/photo2
        - png
        name: doggie
        id: 10
        category:
          name: Dogs
          id: 1
        tags:
        - name: tag1
          id: 1
        - name: tag1
          id: 1
        status: available
      properties:
        id:
          description: Id
          example: 10
          format: int64
          type: integer
        name:
          description: Name
          example: doggie
          type: string
        category:
          $ref: '#/components/schemas/Category'
        photoUrls:
          description: Photo urls.
          example:
          - http://example.com/photo.png
          - http://example.com/photo2
          - png
          items:
            type: string
          type: array
        tags:
          items:
            $ref: '#/components/schemas/Tag'
          type: array
        status:
          description: pet status in the store
          enum:
          - available
          - pending
          - sold
          example: available
          type: string
      required:
      - name
      - photoUrls
      type: object
    ApiResponse:
      example:
        code: 123
        type: responsetype
        message: Message.
      properties:
        code:
          description: Code
          example: 123
          format: int32
          type: integer
        type:
          description: Response type
          example: responsetype
          type: string
        message:
          description: The message.
          example: Message.
          type: string
      type: object
  securitySchemes:
    petstore_auth:
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
      type: oauth2
    api_key:
      in: header
      name: api_key
      type: apiKey